# src/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini Note RESTful API",
        "license": {
            "name": "Apache"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:2880/api/v1",
            "description": "API server development"
        },
        {
            "url": "https://mininote.orlangur.link/api/v1",
            "description": "API server production"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Categories list",
                "description": "View categories information",
                "operationId": "categories.list",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create category",
                "description": "Create category note",
                "operationId": "categories.create",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name category",
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "description": "Parent id child category",
                                        "type": "string"
                                    },
                                    "sort": {
                                        "description": "Sort category",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Read category",
                "description": "Read category note",
                "operationId": "categories.read",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category",
                "description": "Update category note",
                "operationId": "categories.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name category",
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "description": "Parent id child category",
                                        "type": "string"
                                    },
                                    "sort": {
                                        "description": "Sort category",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category",
                "description": "Delete category note",
                "operationId": "categories.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UniversalDTO": {
                "title": "UniversalDTO",
                "required": [],
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "type": "integer",
                                "example": "10"
                            },
                            "message": {
                                "type": "string",
                                "example": "Error message"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}